pyproject.toml
[project]
name = "heartbeatmonitor"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "dishka>=1.9.1",
    "fastapi>=0.135.1",
    "greenlet>=3.3.2",
    "gunicorn>=25.1.0",
    "loguru>=0.7.3",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pyright>=1.1.408",
    "python-dotenv>=1.2.2",
    "ruff>=0.15.5",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
]

[tool.uv]
package = true

[tool.ruff]
line-length = 120
target-version = "py313"
fix = false
exclude = [".venv", "__pycache__", "**/alembic"]

[tool.ruff.lint]
select = ["E", "F", "B", "UP", "I", "SIM", "C4", "RUF", "ANN", "N", "S", "T", "Q", "A", "R", "B9", "W", "PL", "ISC",
    "PIE", "PGH", "TID", "COM", "RET", "PTH", "G", "FA", "ERA", "TRY", "RSE", "SLF", "EXE", "NPY", "PD", "PERF", "PYI",
    "EM", "INP", "PT", "DTZ", "LOG", "BLE", "TCH", "FBT", "ASYNC", "TD"]
ignore = [
    "RUF001",
    "TRY301",
    "BLE001",
    "TD002",
    "TD003",
]
[tool.ruff.lint.flake8-bugbear]
extend-immutable-calls = ["fastapi.Depends", "fastapi.Query", "fastapi.File"]
[tool.ruff.lint.flake8-type-checking]
runtime-evaluated-base-classes = ["pydantic.BaseModel"]

[tool.pyright]
typeCheckingMode = "strict"
reportMissingTypeStubs = "none"
pythonVersion = "3.13"
exclude = [
    ".venv",
    "**/__pycache__",
    "**/alembic",
]
extraPaths = [
    "src"
]
venvPath = "."
venv = ".venv"
